Move UserPresence to ui-common
